LayoutTests/platform/mac/accessibility/aria-liveregions-notifications-always-sent-expected.txt

Test test0test1test2
This tests that ARIA live region notifications are sent even when an assistive technology is not accessing them directly on each update.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS successfullyParsed is true

TEST COMPLETE
